feat: rename module bt convention
