+++ src/hooks/index.ts
@@ -25,11 +25,11 @@ export { default as useExpandableOutcomes } from './useExpandableOutcomes';
 
 export { default as useFantasyTokenTicker } from './useFantasyTokenTicker';
 
+export { default as useMarketColors } from './useMarketColors';
+
 export { default as useMarkets } from './useMarkets';
 
 export { default as useMarketForms } from './useMarketForms';
-
-export { default as useMarketsColors } from './useMarketsColors';
 
 export { default as useMarketPath } from './useMarketPath';
